Reduce permission from "System" to "System::Privileged" to "wx"
"System" label was given broad access to the new "System::Privileged" label.
The full access permitted "System" to fully access files and processes
with "System::Privileged" access, potentially allowing it to gain
CAP_MAC_ADMIN, blocked by onlycap.
One possible way of exploitation would be ptrace from process with "System"
label to process with "System::Privileged". If both processes run as root,
with the same capabilities, ptrace will be permitted. Then the "System"
process is enabled to execute arbitrary code in the context of ptraced
process, circumventing onlycap restriction.

This patch reduces the access to "wx". It will be enough for IPC, which
should be everything needed between "System" and "System::Privileged".

Change-Id: I5259676dba77bbb89d24a37a718b60170f8526f1
Signed-off-by: Rafal Krypa <r.krypa@samsung.com>